# docs/swagger.yaml
definitions:
  models.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: API Documentation for Spotify
  title: Spotify API
  version: 1.0.0
paths:
...
  /addImg:
    post:
      consumes:
      - application/json
      description: Add Thumbnail for Song
      parameters:
      - description: enter Song id and path of thumbnail of song SchemaExample({
        in: body
        name: details
        required: true
        schema:
          example: '{"id":"xyz","img_path":"/"}'
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      tags:
      - Admin
  /addSong:
    post:
      consumes:
      - application/json
      description: Add Song into app
      parameters:
      - description: enter PATH of song SchemaExample({
        in: body
        name: details
        required: true
        schema:
          example: '{"path":"/home/chicmic/Downloads/"}'
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      tags:
      - Admin
...
  /createAlbum:
    post:
      consumes:
      - application/json
      description: Create Album
      parameters:
      - description: enter Song id and album name SchemaExample({
        in: body
        name: details
        required: true
        schema:
          example: '{"song_id":"xyz","album_name":"name"}'
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      tags:
      - Admin
...
swagger: "2.0"